✅ ROOT-DEPLOYMENT: Vite Build direkt ins Repository Root kopiert
## 🎯 Lösung für GitHub Pages Deployment:

**Problem**: GitHub Pages serviert aus Repository Root, nicht aus `/dist`
**Lösung**: Lokales Vite Build direkt ins Root-Verzeichnis kopiert

## 🔄 Geänderte Dateien:

- **games/connect4/index.html**: Vite-prozessierte Version mit korrekten Asset-Links
- **assets/**: Alle CSS/JS Assets aus `/dist/assets` kopiert
- **Alle anderen Spiele**: Ebenfalls mit korrekten Vite-Builds aktualisiert

## 📊 Erwartetes Ergebnis:

- ✅ **Korrekte Asset-Links**: `/LogicCastle/assets/connect4-DG8bBE23.js`
- ✅ **CSS lädt**: `/LogicCastle/assets/connect4-Ddes7ABA.css`
- ✅ **Module Preloads**: Alle Dependencies verfügbar
- ✅ **7x6 Connect4-Spielfeld**: Sollte jetzt korrekt angezeigt werden

+++ game_engine/src/ai/pattern_evaluator.rs
@@ -56,7 +56,7 @@ impl PatternEvaluator {
         // This is much faster than 42 get_cell() calls (6x7)
         for row in 0..6 {
             for col in 0..7 {
-                if let Some(index) = game.geometry().to_index((row as i32, col as i32)) {
+                if let Some(index) = game.geometry().to_index((row, col)) {
                     let cell_value = if yellow_board.get_bit(index) {
                         1u64 // Yellow
                     } else if red_board.get_bit(index) {
